cleaning up board info into it's own file
